+++ tidy3d/plugins/waveguide/__init__.py
@@ -1,0 +1,3 @@
+"""Waveguide utilities module"""
+
+from .rectangular_dielectric import RectangularDielectric